Refactoring. Moved FeignClient to the Common Project.
