+++ Backend/apigateway/src/main/java/com/bikenest/apigateway/ApigatewayApplication.java
@@ -1,17 +1,19 @@
 package com.bikenest.apigateway;
 
+import com.bikenest.common.feignclients.UsermgmtClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.cloud.gateway.route.RouteLocator;
 import org.springframework.cloud.gateway.route.builder.RouteLocatorBuilder;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication(exclude = SecurityAutoConfiguration.class)
-@EnableFeignClients(basePackageClasses = com.bikenest.apigateway.UsermgmtClient.class)
+@ComponentScan({"com.bikenest.common", "com.bikenest.apigateway"})
+@EnableFeignClients(basePackageClasses = UsermgmtClient.class)
 public class ApigatewayApplication {
 
     @Autowired
