+++ Backend/apigateway/src/main/java/com/bikenest/apigateway/AuthenticationFilter.java
@@ -1,11 +1,10 @@
 package com.bikenest.apigateway;
 
+import com.bikenest.common.feignclients.UsermgmtClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.cloud.gateway.filter.GatewayFilter;
 import org.springframework.cloud.gateway.filter.GatewayFilterChain;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ServerWebExchange;
